requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
dateutils==0.6.6
Mastodon.py==1.0.7
nltk==3.2.2
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
six==1.10.0
textblob==0.12.0
